Avoid linking to URLs with 301 redirects
Redirects take time while users are impatient. Depending on the server
performance and realization, a redirect may take about 0,1 to 1 second of
additional loading time. The loading time of a page is a not to be
underestimated ranking factor. Try to avoid 301-redirects (moved permanently)
for internal links as they take time and some link power may be lost.