Use Jekylls default markdown parser
It seems that redcarpet isn't working out of the box with Jekyll 4.0 and I
haven't been able to find the correct configuration conjuration to make it work:

    Markdown processor: "redcarpet" is not a valid Markdown processor.
                        Available processors are: kramdown

      Conversion error: Jekyll::Converters::Markdown encountered an error while converting '_posts/2002-02-20-this-is-a-weblog.markdown':
                        Invalid Markdown processor given: redcarpet
                 ERROR: YOUR SITE COULD NOT BE BUILT:
                        ------------------------------------
                        Invalid Markdown processor given: redcarpet

A cursory glance over some of the posts revealed no apparent issues, although
Kramdown and Redcarpet aren't entirely compatible. We'll fix 'em if we find 'em.