Hardcode robots-tags for category archives
It seems category_archives doesn't pick up the frontmatter variables
from the layout, so our robots meta-tag was blank, which in return
caused the pages to appear in Googles search results.

Ending up on a huge list of entries isn't a great experience, so we'll
reinstate noindex,follow on those pages with this change.